Docs : 로그아웃 기능 문서화
